Use Image component everywhere
