+++ src/@coorpacademy/components/atom/image/style.css
