+++ src/@coorpacademy/components/atom/image/test/fixtures/default.js
@@ -1,6 +1,6 @@
 export default {
   props: {
-    images: {
+    src: {
       '1x': 'https://placehold.it/150x150/1d1d1d',
       '2x': 'https://placehold.it/300x300/1d1d1d'
     }
