+++ src/@coorpacademy/components/atom/image/test/fixtures/empty.js
