+++ src/@coorpacademy/components/atom/image/test/fixtures/simple-src.js
@@ -1,0 +1,5 @@
+export default {
+  props: {
+    src: 'https://placehold.it/150x150/1d1d1d'
+  }
+};
