--- src/@coorpacademy/components/molecule/responsive-image/test/fixtures/simple-src.js
@@ -1,7 +1,0 @@
-export default {
-  props: {
-    image: {
-      src: 'https://placehold.it/150x150/1d1d1d'
-    }
-  }
-};
