+++ src/@coorpacademy/components/molecule/video-player/index.js
@@ -1,6 +1,7 @@
 import {checker, createValidate} from '../../util/validation';
 import style from './style.css';
 import createVideoIframe from '../video-iframe';
+import createImage from '../../atom/image';
 
 const conditions = checker.shape({
   props: checker.shape({
@@ -16,6 +17,7 @@ const conditions = checker.shape({
 
 export default (treant, options = {}) => {
   const VideoIframe = createVideoIframe(treant, options);
+  const Image = createImage(treant, options);
 
   const VideoPlayer = (props, children) => {
     const {h} = treant;
@@ -37,24 +39,28 @@ export default (treant, options = {}) => {
             checked={false}
             className={style.checkbox}
           />
-          <label htmlFor={'toggler'}
-                 className={style.togglerDisplay}
-                 onClick={playVideo}>
-            <img src={image}
-                 className={style.image}
-                 width={width}
-                 height={height}
+          <label
+            htmlFor={'toggler'}
+            className={style.togglerDisplay}
+            onClick={playVideo}
+          >
+            <Image
+              src={image}
+              className={style.image}
+              width={width}
+              height={height}
             />
           </label>
 
           <VideoIframe
-                type={type}
-                id={id}
-                width={width}
-                height={height}
-                frameborder={0}
-                className={style.iframe}
-                allowfullscreen>
+            type={type}
+            id={id}
+            width={width}
+            height={height}
+            frameborder={0}
+            className={style.iframe}
+            allowfullscreen
+          >
           </VideoIframe>
         </div>
       </div>
