+++ src/@coorpacademy/components/organism/discipline-rightaside/index.js
@@ -4,9 +4,8 @@ import getOr from 'lodash/fp/getOr';
 import map from 'lodash/fp/map';
 import style from './style.css';
 import hyperx from 'hyperx';
-import createStarRating from '../../molecule/star-rating';
 import createCatalogCTA from '../../molecule/catalog-cta';
-import createLink from '../../atom/link';
+import createImage from '../../atom/image';
 import HoverFillBehaviour from '../../behaviour/effects/hover-fill';
 
 const conditions = checker.shape({
@@ -41,6 +40,7 @@ export default (treant, options = {}) => {
   const {h} = treant;
   const {translate} = options;
   const CatalogCTA = createCatalogCTA(treant, options);
+  const Image = createImage(treant, options);
 
   const t = stuff => {
     return translate ? translate(stuff) : stuff;
@@ -77,7 +77,7 @@ export default (treant, options = {}) => {
             {authorLabel}
           </div>
           <div className={style.logoContainer}>
-            <img
+            <Image
               className={style.logo}
               src={authorLogo}
             />
